test: Update + Fix Tests
This updates the tests to support the refactored config interface. There are some file
moves to ensure they were picked up by pytests automatic discovery.
