+++ tests/cli/test_config.py
@@ -3,8 +3,7 @@ from unittest import mock
 
 from click.testing import CliRunner
 
-from cally.cli.commands.config import print_service
-from cally.cli.config import CallyConfig
+from cally.cli.commands.config import print_service, print_services
 
 from .. import CallyTestHarness
 
@@ -15,10 +14,20 @@ class ConfigCliTests(CallyTestHarness):
         result = CliRunner().invoke(
             print_service,
             ['--service', 'test-cli', '--environment', 'test'],
-            obj=CallyConfig(config_file='blah.yaml'),
         )
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(
             result.output,
             'ENVIRONMENT: test\nNAME: test-cli\nSTACK_TYPE: CallyStack\n\n',
+        )
+
+    def test_print_services_basic(self):
+        result = CliRunner().invoke(
+            print_services,
+            ['--environment', 'test'],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertEqual(
+            result.output,
+            'ENVIRONMENT: test\n\n',
         )
