+++ tests/cli/test_idp.py
@@ -1,6 +1,6 @@
-from cally import cli
-from cally.cli.config import CallyConfig
 from click.testing import CliRunner
+
+from cally import cli
 
 from .. import CallyIdpTestHarness
 
@@ -11,7 +11,6 @@ class CliTests(CallyIdpTestHarness):
         result = CliRunner().invoke(
             cli.cally,
             ['example', 'hello', 'World'],
-            obj=CallyConfig(config_file='blah.yaml'),
         )
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(result.output, 'Hello World\n')
