+++ tests/cli/test_tf.py
@@ -5,7 +5,6 @@ from unittest import mock
 from click.testing import CliRunner
 
 from cally.cli.commands.tf import tf
-from cally.cli.config import CallyConfig
 
 from .. import CallyTestHarness
 
@@ -17,7 +16,6 @@ class TfTests(CallyTestHarness):
         result = CliRunner().invoke(
             tf,
             ['print', '--service', 'test', '--environment', 'test'],
-            obj=CallyConfig(config_file='blah.yaml'),
         )
         self.assertEqual(result.exit_code, 0)
         testdata = {"backend": {"local": {"path": "state/test/test"}}}
